# webapp/contracts/openapi/bundled.yaml
openapi: 3.1.0
info:
  title: BizSim Edu API
  description: 商域 - 商业模拟教育平台后端 API
  version: 1.0.0
paths:
...
  /api/v1/assignments:
    post:
      tags:
      - 作业
      summary: Create Assignment
      operationId: create_assignment_api_v1_assignments_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: teaching_group_id
        in: query
        required: true
        schema:
          type: integer
          title: Teaching Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AssignmentOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assignments/by-group/{group_id}:
    get:
      tags:
      - 作业
      summary: List Assignments
      operationId: list_assignments_api_v1_assignments_by_group__group_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_AssignmentOut__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assignments/{assignment_id}:
    get:
      tags:
      - 作业
      summary: Get Assignment
      operationId: get_assignment_api_v1_assignments__assignment_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: assignment_id
        in: path
        required: true
        schema:
          type: integer
          title: Assignment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AssignmentOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - 作业
      summary: Update Assignment
      operationId: update_assignment_api_v1_assignments__assignment_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: assignment_id
        in: path
        required: true
        schema:
          type: integer
          title: Assignment Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AssignmentOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - 作业
      summary: Delete Assignment
      operationId: delete_assignment_api_v1_assignments__assignment_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: assignment_id
        in: path
        required: true
        schema:
          type: integer
          title: Assignment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assignments/{assignment_id}/submissions:
    post:
      tags:
      - 作业
      summary: Create Submission
      operationId: create_submission_api_v1_assignments__assignment_id__submissions_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: assignment_id
        in: path
        required: true
        schema:
          type: integer
          title: Assignment Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_SubmissionOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - 作业
      summary: List Submissions
      operationId: list_submissions_api_v1_assignments__assignment_id__submissions_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: assignment_id
        in: path
        required: true
        schema:
          type: integer
          title: Assignment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_SubmissionOut__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assignments/{assignment_id}/submissions/{submission_id}/grade:
    patch:
      tags:
      - 作业
      summary: Grade Submission
      operationId: grade_submission_api_v1_assignments__assignment_id__submissions__submission_id__grade_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: assignment_id
        in: path
        required: true
        schema:
          type: integer
          title: Assignment Id
      - name: submission_id
        in: path
        required: true
        schema:
          type: integer
          title: Submission Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeSubmission'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_SubmissionOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_AssignmentOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/AssignmentOut'
          - type: 'null'
      type: object
      title: ApiResponse[AssignmentOut]
...
    ApiResponse_List_AssignmentOut__:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/AssignmentOut'
            type: array
          - type: 'null'
          title: Data
      type: object
      title: ApiResponse[List[AssignmentOut]]
...
    ApiResponse_List_SubmissionOut__:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/SubmissionOut'
            type: array
          - type: 'null'
          title: Data
      type: object
      title: ApiResponse[List[SubmissionOut]]
...
    ApiResponse_SubmissionOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/SubmissionOut'
          - type: 'null'
      type: object
      title: ApiResponse[SubmissionOut]
...
    AssignmentCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        season_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Season Id
        milestone_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Milestone Id
      type: object
      required:
      - title
      title: AssignmentCreate
    AssignmentOut:
      properties:
        id:
          type: integer
          title: Id
        teaching_group_id:
          type: integer
          title: Teaching Group Id
        season_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Season Id
        milestone_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Milestone Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        created_by:
          type: integer
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        status:
          type: string
          title: Status
        submission_count:
          type: integer
          title: Submission Count
          default: 0
        total_students:
          type: integer
          title: Total Students
          default: 0
      type: object
      required:
      - id
      - teaching_group_id
      - title
      - created_by
      - created_at
      - status
      title: AssignmentOut
    AssignmentUpdate:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        status:
          anyOf:
          - type: string
            pattern: ^(active|closed)$
          - type: 'null'
          title: Status
      type: object
      title: AssignmentUpdate
...
    GradeSubmission:
      properties:
        score:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Score
        feedback:
          anyOf:
          - type: string
          - type: 'null'
          title: Feedback
      type: object
      required:
      - score
      title: GradeSubmission
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SubmissionCreate:
      properties:
        content:
          type: string
          minLength: 1
          title: Content
      type: object
      required:
      - content
      title: SubmissionCreate
    SubmissionOut:
      properties:
        id:
          type: integer
          title: Id
        assignment_id:
          type: integer
          title: Assignment Id
        user_id:
          type: integer
          title: User Id
        username:
          type: string
          title: Username
        content:
          type: string
          title: Content
        submitted_at:
          type: string
          format: date-time
          title: Submitted At
        score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Score
        feedback:
          anyOf:
          - type: string
          - type: 'null'
          title: Feedback
        graded_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Graded At
        graded_by:
          anyOf:
          - type: integer
          - type: 'null'
          title: Graded By
      type: object
      required:
      - id
      - assignment_id
      - user_id
      - username
      - content
      - submitted_at
      title: SubmissionOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login